refactor(Z2): split 3 pre-existing >500 LOC files
rustarr-zha.17: config.rs -> config/{services,mcp,auth}.rs; cli/setup.rs ->
cli/setup/{wizard,plugin}.rs; logging/formatter.rs -> formatter/{render,style}.rs.
Pure mechanical split, public API unchanged via facade re-exports, no mod.rs.
Whole repo now <500 LOC/file. 502 tests, clippy clean.
