+++ src/config/auth.rs
@@ -1,0 +1,76 @@
+//! OAuth / JWT auth sub-config ([`AuthConfig`], [`AuthMode`]) and its defaults.
+
+use serde::{Deserialize, Serialize};
+
+/// OAuth / JWT auth sub-config.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(default)]
+pub struct AuthConfig {
+    pub mode: AuthMode,
+    pub public_url: Option<String>,
+    pub google_client_id: Option<String>,
+    pub google_client_secret: Option<String>,
+    pub admin_email: String,
+    pub allowed_emails: Vec<String>,
+    pub sqlite_path: String,
+    pub key_path: String,
+    pub access_token_ttl_secs: u64,
+    pub refresh_token_ttl_secs: u64,
+    pub auth_code_ttl_secs: u64,
+    pub register_rpm: u32,
+    pub authorize_rpm: u32,
+    pub allowed_client_redirect_uris: Vec<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq, Eq)]
+#[serde(rename_all = "lowercase")]
+pub enum AuthMode {
+    #[default]
+    Bearer,
+    OAuth,
+}
+
+// ── defaults ──────────────────────────────────────────────────────────────────
+
+fn default_auth_sqlite_path() -> String {
+    "/data/auth.db".into()
+}
+fn default_auth_key_path() -> String {
+    "/data/auth-jwt.pem".into()
+}
+fn default_access_token_ttl_secs() -> u64 {
+    3600
+}
+fn default_refresh_token_ttl_secs() -> u64 {
+    86400 * 30
+}
+fn default_auth_code_ttl_secs() -> u64 {
+    300
+}
+fn default_register_rpm() -> u32 {
+    10
+}
+fn default_authorize_rpm() -> u32 {
+    60
+}
+
+impl Default for AuthConfig {
+    fn default() -> Self {
+        Self {
+            mode: AuthMode::default(),
+            public_url: None,
+            google_client_id: None,
+            google_client_secret: None,
+            admin_email: String::new(),
+            allowed_emails: Vec::new(),
+            sqlite_path: default_auth_sqlite_path(),
+            key_path: default_auth_key_path(),
+            access_token_ttl_secs: default_access_token_ttl_secs(),
+            refresh_token_ttl_secs: default_refresh_token_ttl_secs(),
+            auth_code_ttl_secs: default_auth_code_ttl_secs(),
+            register_rpm: default_register_rpm(),
+            authorize_rpm: default_authorize_rpm(),
+            allowed_client_redirect_uris: Vec::new(),
+        }
+    }
+}
