+++ src/config/mcp.rs
@@ -1,0 +1,143 @@
+//! MCP HTTP server configuration ([`McpConfig`]) and the shared `RUSTARR_*`
+//! environment-variable parsing helpers used during [`super::Config::load`].
+
+use serde::{Deserialize, Serialize};
+
+use super::AuthConfig;
+
+/// MCP HTTP server configuration.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(default)]
+pub struct McpConfig {
+    /// Bind host (RUSTARR_MCP_HOST). Default: `127.0.0.1` (loopback).
+    /// Set to `0.0.0.0` to listen on all interfaces — requires auth configured.
+    #[serde(default = "default_mcp_host")]
+    pub host: String,
+    /// Bind port (RUSTARR_MCP_PORT). Default: `40070`.
+    #[serde(default = "default_mcp_port")]
+    pub port: u16,
+    /// MCP server name advertised to clients (RUSTARR_MCP_SERVER_NAME).
+    #[serde(default = "default_server_name")]
+    pub server_name: String,
+    /// Disable auth entirely — only safe when bound to loopback (RUSTARR_MCP_NO_AUTH).
+    pub no_auth: bool,
+    /// Allow unauthenticated access on non-loopback when behind a trusted reverse proxy
+    /// that enforces its own auth (RUSTARR_NOAUTH). Loaded here so it participates in
+    /// typed config rather than being a raw env read at call sites.
+    pub trusted_gateway: bool,
+    /// Static bearer token for simple auth (RUSTARR_MCP_TOKEN).
+    pub api_token: Option<String>,
+    /// Additional allowed Host header values (comma-separated in env).
+    pub allowed_hosts: Vec<String>,
+    /// Additional allowed CORS origins (comma-separated in env).
+    pub allowed_origins: Vec<String>,
+    /// OAuth sub-config (nested under `[mcp.auth]` in config.toml).
+    pub auth: AuthConfig,
+}
+
+impl McpConfig {
+    pub fn bind_addr(&self) -> String {
+        format!("{}:{}", self.host, self.port)
+    }
+
+    /// Return true if the configured bind host resolves to a loopback address.
+    ///
+    /// Uses `IpAddr::is_loopback()` for numeric addresses. Accepts "localhost"
+    /// as a canonical loopback hostname. Any other hostname or parse failure is
+    /// treated as non-loopback — callers must not assume safety in that case.
+    pub fn is_loopback(&self) -> bool {
+        let host = &self.host;
+        // Match "localhost" literal and numeric loopback addresses.
+        // Strip bracket notation ([::1]) before parsing so IPv6 loopback works.
+        host == "localhost"
+            || host
+                .trim_start_matches('[')
+                .trim_end_matches(']')
+                .parse::<std::net::IpAddr>()
+                .map(|ip| ip.is_loopback())
+                .unwrap_or(false)
+    }
+}
+
+// ── defaults ──────────────────────────────────────────────────────────────────
+
+fn default_mcp_host() -> String {
+    // Default to loopback for safety. Operators who need external access must
+    // explicitly set RUSTARR_MCP_HOST=0.0.0.0 (and configure auth).
+    "127.0.0.1".into()
+}
+pub(super) fn default_mcp_port() -> u16 {
+    40070
+}
+fn default_server_name() -> String {
+    "rustarr-mcp".into()
+}
+
+impl Default for McpConfig {
+    fn default() -> Self {
+        Self {
+            host: default_mcp_host(),
+            port: default_mcp_port(),
+            server_name: default_server_name(),
+            no_auth: false,
+            trusted_gateway: false,
+            api_token: None,
+            allowed_hosts: Vec::new(),
+            allowed_origins: Vec::new(),
+            auth: AuthConfig::default(),
+        }
+    }
+}
+
+// ── env helpers ───────────────────────────────────────────────────────────────
+
+pub(super) fn env_str(key: &str, target: &mut String) {
+    if let Ok(v) = std::env::var(key) {
+        if !v.is_empty() {
+            *target = v;
+        }
+    }
+}
+
+pub(super) fn env_opt_str(key: &str, target: &mut Option<String>) {
+    if let Ok(v) = std::env::var(key) {
+        if !v.is_empty() {
+            *target = Some(v);
+        }
+    }
+}
+
+pub(super) fn env_parse<T: std::str::FromStr>(key: &str, target: &mut T) -> anyhow::Result<()> {
+    if let Ok(v) = std::env::var(key) {
+        if !v.is_empty() {
+            *target = v
+                .parse()
+                .map_err(|_| anyhow::anyhow!("{key}: invalid value {v:?}"))?;
+        }
+    }
+    Ok(())
+}
+
+pub(super) fn env_bool(key: &str, target: &mut bool) -> anyhow::Result<()> {
+    if let Ok(v) = std::env::var(key) {
+        match v.to_lowercase().as_str() {
+            "1" | "true" | "yes" => *target = true,
+            "0" | "false" | "no" => *target = false,
+            other => anyhow::bail!("{key}: expected bool, got {other:?}"),
+        }
+    }
+    Ok(())
+}
+
+pub(super) fn env_list(key: &str, target: &mut Vec<String>) {
+    if let Ok(v) = std::env::var(key) {
+        let items: Vec<String> = v
+            .split(',')
+            .map(|s| s.trim().to_string())
+            .filter(|s| !s.is_empty())
+            .collect();
+        if !items.is_empty() {
+            *target = items;
+        }
+    }
+}
